requirements.txt
openai
langfuse
python-dotenv
requests